move attachments logic to own component class
