+++ app/components/process/modals.hbs
@@ -80,41 +80,3 @@
     </AuButtonGroup>
   </:footer>
 </AuModal>
-
-<AuModal @modalOpen={{@addModalOpened}} @closeModal={{@closeAddModal}}>
-  <:title>Voeg bijlage toe</:title>
-  <:body>
-    <CustomFileUpload
-      @title="Bestand selecteren"
-      @helpTextDragDrop="Bestand wordt meteen toegevoegd aan bibliotheek na selectie"
-      @forbidden=".bpmn,.vsdx"
-      @multiple="true"
-      @updateProcess={{@addFileToProcess}}
-      @onFinishUpload={{@attachmentsUploaded}}
-    />
-  </:body>
-  <:footer>
-    <AuButtonGroup class="au-u-flex--end">
-      <AuButton @skin="naked" {{on "click" @closeAddModal}}>Annuleer</AuButton>
-    </AuButtonGroup>
-  </:footer>
-</AuModal>
-
-<AuModal @modalOpen={{@deleteModalOpened}} @closeModal={{@closeDeleteModal}}>
-  <:title>Verwijder {{@fileToDelete.name}}</:title>
-  <:body>Ben je zeker dat je dit bestand uit de bibliotheek wil verwijderen?</:body>
-  <:footer>
-    <AuButtonGroup class="au-u-flex--end">
-      <AuButton
-        @skin="secondary"
-        {{on "click" @closeDeleteModal}}
-      >Annuleer</AuButton>
-      <AuButton
-        @skin="primary"
-        @icon="bin"
-        @alert="true"
-        {{on "click" (perform @deleteFile)}}
-      >Verwijder</AuButton>
-    </AuButtonGroup>
-  </:footer>
-</AuModal>
